refactor: Выносит работу c api localStorage в отдельный модуль
