+++ src/store/assync-action.ts
@@ -3,6 +3,7 @@ import { Offers, OfferServer } from './../types/offer';
 import { AuthorisationData } from '../types/authorisation';
 import { requireAuthorization, requireLogout } from './app-user/app-user';
 import { setDataLoaded, setOffers } from './action';
+import { setToken } from '../services/token';
 import { adaptOfferToClient } from '../util';
 import { ApiRoute, AuthorisationStatus } from './../const';
 
@@ -19,7 +20,7 @@ export const loadOffers =
 
 export const authorise = (authorisationData: AuthorisationData): ThunkCreatorResult => (dispatch, _getState, api) => {
   api.post(ApiRoute.Login, authorisationData)
-    .then((response) => localStorage.setItem('X-Token', `${response.data.token}`))
+    .then((response) => setToken(response.data.token))
     .then(() => dispatch(requireAuthorization(AuthorisationStatus.Auth)))
     .then(() => dispatch(requireLogout()));
 };
